+++ src/App/Shared/GoldEx.Shared/DTOs/Customers/UpdateCustomerRequest.cs
@@ -7,6 +7,6 @@ public record UpdateCustomerRequest(Guid Id,
     string NationalId,
     string PhoneNumber,
     string? Address,
-    double? CreditLimit,
+    decimal? CreditLimit,
     UnitType? CreditLimitUnit,
     CustomerType CustomerType);
